Debugging image saving

+++ nvs_from_video/run_eval.py
@@ -30,7 +30,7 @@ def render_colmap_dir_splat(modelparams, iteration, pipelineparams):
         assert(os.path.exists(output_path))
 
         for camera in test_scene.getTrainCameras():
-            rendering = np.array(render_gaussians(camera, gaussians, pipelineparams, background)["render"].cpu().detach())
+            rendering = np.array(render_gaussians(camera, gaussians, pipelineparams, background)["render"].cpu().detach(), dtype=np.int8)
             print(rendering)
             print(rendering.shape)
             filename = camera.image_name + ".jpg"
